requirements.txt
fastapi
uvicorn
httpx
torch
ultralytics
opencv-python
numpy
torchvision
python-multipart
pydantic
pillow
google-cloud-storage
google-generativeai
category_encoders
scikit-learn
pandas
joblib
python-dotenv
